# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: ExpertMatch API
  version: 1.0.0
  description: |
    GraphRAG-powered expert discovery and matching service for RFPs and team formation.
    This API enables automatic client generation for frontend applications using OpenAPI generators.
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: ""
    description: Current Server (auto-detected)
  - url: http://localhost:8093
    description: Local Development Server
  - url: http://192.168.0.73:8093
    description: Remote Development Server
  - url: https://api.expertmatch.com
    description: Production Server

# Security: Authentication and authorization are handled by Spring Gateway.
# Spring Gateway validates JWT tokens and populates user information in HTTP headers:
# - X-User-Id: User identifier (required)
# - X-User-Roles: Comma-separated list of roles (optional, e.g., "ROLE_USER,ROLE_ADMIN")
# - X-User-Email: User email address (optional)
...
components:
  parameters:
    X-User-Id:
      name: X-User-Id
      in: header
      required: false
      description: User identifier populated by Spring Gateway after authentication. If not provided, anonymous user will be used.
      schema:
        type: string
        default: "anonymous-user"
        example: "user-123"
    X-User-Roles:
      name: X-User-Roles
      in: header
      required: false
      description: Comma-separated list of user roles populated by Spring Gateway
      schema:
        type: string
        example: "ROLE_USER,ROLE_ADMIN"
    X-User-Email:
      name: X-User-Email
      in: header
      required: false
      description: User email address populated by Spring Gateway
      schema:
        type: string
        example: "user@example.com"
  schemas:
...
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Natural language query describing expert requirements
          example: "Looking for experts in Java, Spring Boot, and AWS"
          maxLength: 5000
        chatId:
          type: string
          description: Optional chat ID for conversation context (24-character hexadecimal string)
          example: "507f1f77bcf86cd799439011"
          pattern: "^[0-9a-fA-F]{24}$"
          minLength: 24
          maxLength: 24
        options:
          $ref: '#/components/schemas/QueryOptions'
    QueryOptions:
      type: object
      properties:
        maxResults:
          type: integer
          description: Maximum number of results to return (1-100)
          example: 10
          minimum: 1
          maximum: 100
          default: 10
        minConfidence:
          type: number
          format: double
          description: Minimum confidence score (0.0-1.0)
          example: 0.7
          minimum: 0.0
          maximum: 1.0
          default: 0.7
        includeSources:
          type: boolean
          description: Include source citations in response
          example: true
          default: true
        includeEntities:
          type: boolean
          description: Include extracted entities in response
          example: true
          default: true
        rerank:
          type: boolean
          description: Enable semantic reranking
          example: true
          default: true
        deepResearch:
          type: boolean
          description: Enable deep research SGR pattern
          example: false
          default: false
        useCascadePattern:
          type: boolean
          description: Enable Cascade pattern for structured expert evaluation (requires exactly 1 expert result)
          example: false
          default: false
        useRoutingPattern:
          type: boolean
          description: Enable Routing pattern for LLM-based query classification
          example: false
          default: false
        useCyclePattern:
          type: boolean
          description: Enable Cycle pattern for multiple expert evaluations (requires multiple expert results)
          example: false
          default: false
        includeExecutionTrace:
          type: boolean
          description: Include execution trace in response (step-by-step processing details with LLM models and token usage)
          example: false
          default: false
...
    QueryResponse:
      type: object
      required:
        - answer
        - confidence
        - queryId
      properties:
        answer:
          type: string
          description: Natural language answer with expert recommendations
          example: "Based on your requirements, I found several experts..."
        experts:
          type: array
          description: List of matched experts with detailed information
          items:
            $ref: '#/components/schemas/ExpertMatch'
        sources:
          type: array
          description: Source citations used in the answer
          items:
            $ref: '#/components/schemas/Source'
        entities:
          type: array
          description: Extracted entities from the query
          items:
            $ref: '#/components/schemas/Entity'
        confidence:
          type: number
          format: double
          description: Confidence score for the response (0.0-1.0)
          example: 0.85
          minimum: 0.0
          maximum: 1.0
        queryId:
          type: string
          description: Unique query ID
          example: "507f1f77bcf86cd799439012"
        chatId:
          type: string
          description: Chat ID used for context
          example: "507f1f77bcf86cd799439011"
        messageId:
          type: string
          description: Message ID for this response
          example: "507f1f77bcf86cd799439013"
        processingTimeMs:
          type: integer
          format: int64
          description: Processing time in milliseconds
          example: 1234
        summary:
          $ref: '#/components/schemas/MatchSummary'
        executionTrace:
          $ref: '#/components/schemas/ExecutionTrace'
          description: Step-by-step execution trace (only included if includeExecutionTrace=true)
    ExpertMatch:
      type: object
      properties:
        id:
          type: string
          description: Expert ID
        name:
          type: string
          description: Expert name
        email:
          type: string
          description: Expert email
        seniority:
          type: string
          description: Seniority level (e.g., A3, A4, A5)
        language:
          $ref: '#/components/schemas/LanguageProficiency'
        skillMatch:
          $ref: '#/components/schemas/SkillMatch'
        matchedSkills:
          $ref: '#/components/schemas/MatchedSkills'
        relevantProjects:
          type: array
          description: Relevant projects
          items:
            $ref: '#/components/schemas/RelevantProject'
        experience:
          $ref: '#/components/schemas/Experience'
        relevanceScore:
          type: number
          format: double
          description: Relevance score (0-1)
        availability:
          type: string
          description: Availability status
    SkillMatch:
      type: object
      properties:
        mustHaveMatched:
          type: integer
          description: Number of must-have skills matched
        mustHaveTotal:
          type: integer
          description: Total number of must-have skills required
        niceToHaveMatched:
          type: integer
          description: Number of nice-to-have skills matched
        niceToHaveTotal:
          type: integer
          description: Total number of nice-to-have skills
        matchScore:
          type: number
          format: double
          description: Overall match score (0-1)
    MatchedSkills:
      type: object
      properties:
        mustHave:
          type: array
          description: List of matched must-have skills
          items:
            type: string
        niceToHave:
          type: array
          description: List of matched nice-to-have skills
          items:
            type: string
    RelevantProject:
      type: object
      properties:
        name:
          type: string
          description: Project name
        technologies:
          type: array
          description: Technologies used in project
          items:
            type: string
        role:
          type: string
          description: Role in project
        duration:
          type: string
          description: Project duration
    Experience:
      type: object
      properties:
        etlPipelines:
          type: boolean
          description: Has ETL pipeline experience
        highPerformanceServices:
          type: boolean
          description: Has high-performance services experience
        systemArchitecture:
          type: boolean
          description: Has system architecture experience
        monitoring:
          type: boolean
          description: Has monitoring/observability experience
        onCall:
          type: boolean
          description: Has on-call rotation experience
    LanguageProficiency:
      type: object
      properties:
        english:
          type: string
          description: English proficiency level (e.g., B2+)
    MatchSummary:
      type: object
      properties:
        totalExpertsFound:
          type: integer
          description: Total number of experts found
        perfectMatches:
          type: integer
          description: Number of perfect matches (matchScore >= 0.9)
        goodMatches:
          type: integer
          description: Number of good matches (0.7 <= matchScore < 0.9)
        partialMatches:
          type: integer
          description: Number of partial matches (matchScore < 0.7)
    ExecutionTrace:
      type: object
      description: Step-by-step execution trace showing internal processing details
      properties:
        steps:
          type: array
          description: List of execution steps
          items:
            $ref: '#/components/schemas/ExecutionStep'
        totalDurationMs:
          type: integer
          format: int64
          description: Total processing duration in milliseconds
          example: 2402
        totalTokenUsage:
          $ref: '#/components/schemas/TokenUsage'
          description: Aggregated token usage across all LLM steps
    ExecutionStep:
      type: object
      description: Individual execution step with tracking information
      properties:
        name:
          type: string
          description: Step name
          example: "Parse Query"
        service:
          type: string
          description: Service that executed the step
          example: "QueryParser"
        method:
          type: string
          description: Method called
          example: "parse"
        durationMs:
          type: integer
          format: int64
          description: Duration in milliseconds
          example: 234
        status:
          type: string
          enum: [ SUCCESS, FAILED, SKIPPED ]
          description: Step status
          example: "SUCCESS"
        inputSummary:
          type: string
          description: Input summary
          example: "Query: Find Java experts"
        outputSummary:
          type: string
          description: Output summary
          example: "Intent: expert_search, Skills: [Java]"
        llmModel:
          type: string
          nullable: true
          description: LLM model used (null if no LLM)
          example: "OllamaChatModel (qwen3:4b-instruct-2507-q4_K_M)"
        tokenUsage:
          $ref: '#/components/schemas/TokenUsage'
          nullable: true
          description: Token usage (null if no LLM or usage not available)
    TokenUsage:
      type: object
      description: Token usage information for LLM calls
      properties:
        inputTokens:
          type: integer
          nullable: true
          description: Input/prompt tokens
          example: 45
        outputTokens:
          type: integer
          nullable: true
          description: Output/completion tokens
          example: 12
        totalTokens:
          type: integer
          nullable: true
          description: Total tokens (input + output)
          example: 57
    Source:
      type: object
      properties:
        type:
          type: string
          enum: [ expert, project, document ]
          description: Source type
        id:
          type: string
          description: Source ID
        title:
          type: string
          description: Source title
        relevanceScore:
          type: number
          format: double
          description: Relevance score
        metadata:
          type: object
          description: Additional metadata
          additionalProperties: true
    Entity:
      type: object
      properties:
        type:
          type: string
          enum: [ person, organization, technology, project, domain ]
          description: Entity type
        name:
          type: string
          description: Entity name
        id:
          type: string
          description: Entity ID

    # Chat-related schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp
        path:
          type: string
          description: API path that caused the error
        details:
          type: object
          description: Additional error details
          additionalProperties: true
paths:
...
  /query:
    post:
      tags:
        - Query
      summary: Process natural language query
      description: |
        Processes a natural language query to discover experts matching project requirements.
        Uses hybrid GraphRAG retrieval combining vector search, graph traversal, and keyword search.
        If chatId is not provided, uses or creates the user's default chat for conversation context.
      operationId: processQuery
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
            example:
              query: "Looking for experts in Java, Spring Boot, and AWS"
              chatId: "507f1f77bcf86cd799439011"
              options:
                maxResults: 10
                minConfidence: 0.7
                includeSources: true
                includeEntities: true
                rerank: true
                deepResearch: false
                useCascadePattern: false
                useRoutingPattern: false
                useCyclePattern: false
                includeExecutionTrace: false
      responses:
        '200':
          description: Query processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Invalid request format or validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Chat access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: LLM or embedding service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /query-stream:
    post:
      tags:
        - Query Streaming
      summary: Process query with streaming response
      description: |
        Processes a natural language query and streams the response using Server-Sent Events (SSE).
        Emits events at each processing stage: parsing, retrieving, reranking, generating, and complete.
        If chatId is not provided, uses or creates the user's default chat for conversation context.
      operationId: processQueryStream
      parameters:
        - $ref: '#/components/parameters/X-User-Id'
        - $ref: '#/components/parameters/X-User-Roles'
        - $ref: '#/components/parameters/X-User-Email'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
            example:
              query: "Looking for experts in Java, Spring Boot, and AWS"
              chatId: "507f1f77bcf86cd799439011"
              options:
                maxResults: 10
                minConfidence: 0.7
                includeSources: true
                includeEntities: true
                rerank: true
                deepResearch: false
                useCascadePattern: false
                useRoutingPattern: false
                useCyclePattern: false
                includeExecutionTrace: false
      responses:
        '200':
          description: Query processing stream
          content:
            text/event-stream:
              schema:
                type: string
                format: binary
                description: Server-Sent Events stream with query processing updates
        '400':
          description: Invalid request format or validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Chat access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: LLM or embedding service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...